pipeline/edinet_daily_job/requirements.txt
pandas
dotenv
google-cloud-bigquery
